fix order

+++ server/controllers/orderControllers.js
@@ -1,27 +1,30 @@
-const Order = require('../models/Order')
-const Cart = require('../models/Cart')
-const { getPopulatedCart } = require('../utils/cartUtils')
+const { Order } = require('../models/Order')
+const { Cart } = require('../models/Cart')
+const { User } = require('../models/User')
 
-module.exports.get_user_orders = async (req, res) => {
+const getUserOrders = async (req, res) => {
   const userId = req.params.userId
   try {
-    const orders = await Order.find({ userId })
-    res.status(200).json(orders)
+    const user = await User.findOne({ _id: userId })
+    if (user) {
+      const orders = await Order.find({ userId })
+      delete user._doc.password
+      res.status(200).json({ user, ordersNo: orders.length, orders })
+    } else {
+      res.status(404).send({ message: 'User not found' })
+    }
   } catch (err) {
     console.log(err)
+    return res.status(500).json({ message: 'Internal server error.' })
   }
 }
 
-module.exports.create_order = async (req, res) => {
-  console.log('req', req.body)
+const createOrder = async (req, res) => {
   const userId = res.locals.user._id
   const { shippingAddress } = req.body
-  //   console.log(shippingAddress)
   try {
     const cart = await Cart.findOne({ userId })
-    if (!cart) throw Error('Your cart is empty.')
-    const populatedCart = await getPopulatedCart(cart._id)
-    //   res.json(populatedCart)
+    if (!cart) return res.status(404).json('Your cart is empty.')
 
     const { items, subTotal } = cart
     const order = await Order.create({
@@ -31,43 +34,64 @@ module.exports.create_order = async (req, res) => {
       subTotal,
     })
     console.log(order)
-    //delete cart
-    const ackDeletedCart = await Cart.deleteOne({ userId })
-    console.log(ackDeletedCart)
+    // delete cart
+    const result = await Cart.deleteOne({ userId })
+    if (!result.acknowledged) {
+      return res.status(500).json({ message: 'Internal server error.' })
+    }
     res.status(201).json(order)
   } catch (err) {
-    if (err.message === 'Your cart is empty.') return res.json(err.message)
     console.log(err)
+    return res.status(500).json({ message: 'Internal server error.' })
   }
 }
 
-module.exports.delete_order = async (req, res) => {
+const deleteOrder = async (req, res) => {
   const id = req.params.id
 
   try {
-    const ackDeleteOrder = await Order.deleteOne({ _id: id })
-    console.log(ackDeleteOrder)
-    res.status(204).json({ message: 'order deleted successfully.' })
+    const order = await Order.findById(id)
+    if (!order) return res.status(404).json('Order not found.')
+
+    const result = await Order.deleteOne({ _id: id })
+    if (!result.acknowledged) {
+      return res.status(500).json({ message: 'Internal server error.' })
+    }
+    console.log(result)
+    res.status(204).json({ message: 'Order deleted successfully.' })
   } catch (err) {
     console.log(err)
+    return res.status(500).json({ message: 'Internal server error.' })
   }
 }
 
-module.exports.get_my_orders = async (req, res) => {
+const getMyOrders = async (req, res) => {
   const userId = res.locals.user._id
   try {
     const orders = await Order.find({ userId })
-    res.json(orders)
+    if (!orders) return res.status(404).json('Orders not found.')
+    res.json({ ordersNo: orders.length, orders })
   } catch (err) {
     console.log(err)
+    return res.status(500).json({ message: 'Internal server error.' })
   }
 }
-module.exports.get_myorder_by_id = async (req, res) => {
+const getMyorderById = async (req, res) => {
   const id = req.params.id
   try {
-    const order = await Order.findOne({ _id: id })
+    const order = await Order.findById(id)
+    if (!order) return res.status(404).json('Order not found.')
     res.json(order)
   } catch (err) {
     console.log(err)
+    return res.status(500).json({ message: 'Internal server error.' })
   }
+}
+
+module.exports = {
+  getUserOrders,
+  createOrder,
+  deleteOrder,
+  getMyOrders,
+  getMyorderById,
 }
